Added images locally + working on the panel menu
I've created my own workspace to work on the menu for the top of the page. It's design needs to be worked on more, and so I have made it in a different folder instead of implementing changes in the main code.

Images have been added to the project to have easy access to files without having to rely on access online.