Shop page for phone, Subscribe form adjustments, NAV.css document
Compared the figma file with my current setup and had to make some changes as the footer form I had made did not adhere to the form seen on Figma.

Shop page for phone has been worked on and corrected.

Created a nav.css document as I want to work on that part next, stopped to make a new branch for that part.