NAV menu
Nav menu has been fixed so it takes up the correct amount of space for pc screens. I closed the project, then remembered that I need to add phonescreen settings. That will be the next task, along with some simple Javascript to allow you to click out of the menu when you click outside the menu area.